zstd angefangen...
	geändert:       Makefile
	geändert:       lz4mt.h
	geändert:       lz5mt.h
	geändert:       threading.h
	geändert:       zstdmt.c
	geändert:       zstdmt.h
	geändert:       zstdmt_compress.c
	geändert:       zstdmt_decompress.c

+++ zstdmt_compress.c
@@ -16,6 +16,7 @@
 #define ZSTD_STATIC_LINKING_ONLY
 #include "zstd.h"
 
+#include "mem.h"
 #include "threading.h"
 #include "list.h"
 #include "zstdmt.h"
@@ -390,7 +391,7 @@ static void *pt_compress(void *arg)
 	return (void *)-1;
 }
 
-int ZSTDMT_CompressCCtx(ZSTDMT_CCtx * ctx, ZSTDMT_RdWr_t * rdwr)
+size_t ZSTDMT_CompressCCtx(ZSTDMT_CCtx * ctx, ZSTDMT_RdWr_t * rdwr)
 {
 	int t;
